moved config to top-level
