+++ doc/source/config.rst
@@ -1,0 +1,4 @@
+.. meta::
+   :description: User configuration utilities.
+
+.. automodule:: skbio._config
