+++ skbio/io/format/tests/test_ordination.py
@@ -8,7 +8,6 @@
 
 import io
 from unittest import TestCase, main
-import itertools
 
 import numpy as np
 import pandas as pd
@@ -20,7 +19,7 @@ from skbio.io.format.ordination import (
     _ordination_to_ordination_results, _ordination_results_to_ordination,
     _ordination_sniffer)
 from skbio.util import get_data_path, assert_ordination_results_equal, assert_ordination_results_equal_np
-from skbio.util.config._config import set_config
+from skbio._config import set_config
 
 
 class OrdinationTestData(TestCase):
