+++ skbio/stats/ordination/tests/test_correspondence_analysis.py
@@ -13,7 +13,7 @@ import numpy.testing as npt
 import pandas as pd
 from scipy.spatial.distance import pdist
 
-from skbio import set_config
+from skbio._config import set_config
 from skbio import OrdinationResults
 from skbio.stats.ordination import ca
 from skbio.util import (
