+++ skbio/util/__init__.py
@@ -8,19 +8,6 @@ scikit-bio, as well as various utility functionality, including I/O and
 unit-testing convenience functions.
 
 
-User configuration utilities
-----------------------------
-
-Configuration options for functional input and output.
-
-.. autosummary::
-   :toctree: generated/
-
-   config
-   get_config
-   set_config
-
-
 Testing utilities
 -----------------
 
@@ -119,7 +106,6 @@ from ._decorator import (
 )
 from ._optionals import get_package
 from ._plotting import PlottableMixin
-from .config._config import get_config, set_config
 from ._types import DataTable
 
 __all__ = [
@@ -141,7 +127,5 @@ __all__ = [
     "aliased",
     "register_aliases",
     "params_aliased",
-    "get_config",
-    "set_config",
     "DataTable",
 ]
